meta/plugins/requirements.txt
python-magic ~= 0.4.27
git+https://github.com/cute-engineering/cutekit.git@0.8-dev
git+https://github.com/cute-engineering/chatty
